Add files via upload
This application is called the Sentence Maker. it asks the user to say something repeatedly until the enter \end. All the sentences that they type till the \end will be displayed continuously separated by . and capitalising the first letter of the sentence and its adds a ? if the sentence is a question